refactored ignore.rules package and classes.
