+++ modules/swagger-codegen/src/main/java/io/swagger/codegen/v3/ignore/rules/IgnoreLineParser.java
@@ -1,4 +1,4 @@
-package io.swagger.codegen.ignore.rules;
+package io.swagger.codegen.v3.ignore.rules;
 
 import java.util.ArrayList;
 import java.util.List;
