+++ modules/swagger-codegen/src/test/java/io/swagger/codegen/v3/ignore/CodegenIgnoreProcessorTest.java
@@ -1,4 +1,4 @@
-package io.swagger.codegen.ignore;
+package io.swagger.codegen.v3.ignore;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.SystemUtils;
