+++ modules/swagger-codegen/src/test/java/io/swagger/codegen/v3/ignore/rules/FileRuleTest.java
@@ -1,5 +1,9 @@
-package io.swagger.codegen.ignore.rules;
+package io.swagger.codegen.v3.ignore.rules;
 
+import io.swagger.codegen.v3.ignore.rules.FileRule;
+import io.swagger.codegen.v3.ignore.rules.IgnoreLineParser;
+import io.swagger.codegen.v3.ignore.rules.Part;
+import io.swagger.codegen.v3.ignore.rules.Rule;
 import org.testng.annotations.Test;
 
 import java.util.Arrays;
