+++ modules/swagger-codegen/src/test/java/io/swagger/codegen/v3/ignore/rules/RootedFileRuleTest.java
@@ -1,5 +1,9 @@
-package io.swagger.codegen.ignore.rules;
+package io.swagger.codegen.v3.ignore.rules;
 
+import io.swagger.codegen.v3.ignore.rules.IgnoreLineParser;
+import io.swagger.codegen.v3.ignore.rules.Part;
+import io.swagger.codegen.v3.ignore.rules.RootedFileRule;
+import io.swagger.codegen.v3.ignore.rules.Rule;
 import org.testng.annotations.Test;
 
 import java.util.Arrays;
